+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/model/ProductIngestResponse.java
@@ -1,10 +1,13 @@
 package com.backbase.stream.compositions.product.core.model;
 
-import lombok.*;
 import com.backbase.stream.legalentity.model.ProductGroup;
-
 import java.util.List;
 import java.util.Map;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.With;
 
 @Setter
 @Getter
@@ -34,4 +37,6 @@ public class ProductIngestResponse {
     @With
     private Map<String, String> additions;
     private String source;
+    private Boolean transactionChainEnabledFromRequest;
+    private Boolean paymentOrderChainEnabledFromRequest;
 }