+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/service/impl/ProductPostIngestionServiceImpl.java
@@ -69,7 +69,11 @@ public class ProductPostIngestionServiceImpl implements ProductPostIngestionServ
     private Mono<ProductIngestResponse> processTransactionChains(ProductIngestResponse res) {
         Mono<ProductIngestResponse> transactionChainMono;
 
-        if (!config.isTransactionChainEnabled()) {
+        boolean isTransactionChainEnabled = res.getTransactionChainEnabledFromRequest() == null
+            ? config.isTransactionChainEnabled()
+            : res.getTransactionChainEnabledFromRequest();
+
+        if (!isTransactionChainEnabled) {
             log.debug("Transaction Chain is disabled");
             transactionChainMono = Mono.just(res);
         } else if (config.isTransactionChainAsync()) {
@@ -84,7 +88,11 @@ public class ProductPostIngestionServiceImpl implements ProductPostIngestionServ
     private Mono<ProductIngestResponse> processPaymentOrderChains(ProductIngestResponse res) {
         Mono<ProductIngestResponse> paymentOrderChainMono;
 
-        if (!config.isPaymentOrderChainEnabled()) {
+        boolean isPaymentOrderChainEnabled = res.getPaymentOrderChainEnabledFromRequest() == null
+            ? config.isPaymentOrderChainEnabled()
+            : res.getPaymentOrderChainEnabledFromRequest();
+
+        if (!isPaymentOrderChainEnabled) {
             log.debug("Payment Order Chain is disabled");
             paymentOrderChainMono = Mono.just(res);
         } else if (config.isPaymentOrderChainAsync()) {