+++ stream-portfolio/portfolio-core/src/test/java/com/backbase/stream/portfolio/configuration/PortfolioApiConfigurationTest.java
@@ -1,8 +1,8 @@
 package com.backbase.stream.portfolio.configuration;
 
-import org.junit.jupiter.api.Test;
-
 import com.backbase.portfolio.integration.api.service.ApiClient;
+import com.backbase.stream.clients.config.PortfolioApiConfiguration;
+import org.junit.jupiter.api.Test;
 
 class PortfolioApiConfigurationTest extends BaseApiConfigurationTest<PortfolioApiConfiguration, ApiClient> {