+++ stream-product/product-core/src/test/java/com/backbase/stream/product/mapping/ProductMapperTest.java
@@ -327,7 +327,7 @@ public class ProductMapperTest {
     }
 
     @Test
-    public void mapCurrentInvestment() {
+    void mapCurrentInvestment() {
         BigDecimal source = new BigDecimal("12345.69");
         CurrentInvestment target = productMapper.mapCurrentInvestment(source);
         Assertions.assertEquals(source, target.getAmount());